// frontend/src/interfaces/enums/session.status.enum.ts
export enum SessionStatusEnum {
	PENDING = "pending",
	ACCEPTED = "accepted",
	REJECTED = "rejected",
	CANCELED = "canceled",
	UPCOMING = "upcoming",
	COMPLETED = "completed",
	EXPIRED = "expired",
	APPROVED = "approved",
}

export enum SessionPaymentStatusEnum {
	PENDING = "pending",
	COMPLETED = "completed",
	FAILED = "failed",
}